Moved ARC-170 config
