+++ BNA_KC_Vehicles/Aviation/CfgPatches.hpp
@@ -28,7 +28,6 @@ class CfgPatches
             "BNA_KC_LAATi_MK2_Lamps",
             "BNA_KC_LAATc",
             "BNA_KC_RepubTransport",
-            "BNA_KC_ARC_170",
             "BNA_KC_Galaxy_Gunship",
             "BNA_KC_HAG_Base",
             "BNA_KC_Galaxy_Transport_Vehicle",
